sqlite: rename parse -> jsonParseStream
