+++ lib/jsonParseStream.js
